if best tile not found, send null in ClientGameRunner
