+++ src/core/pathfinding/MiniAStar.ts
@@ -2,7 +2,6 @@ import { Cell } from "../game/Game";
 import { GameMap, TileRef } from "../game/GameMap";
 import { AStar, PathFindResultType } from "./AStar";
 import { SerialAStar } from "./SerialAStar";
-import { SerialAStarOld } from "./SerialAStarOld";
 
 export class MiniAStar implements AStar {
   private aStar: AStar;
@@ -14,40 +13,27 @@ export class MiniAStar implements AStar {
     private dst: TileRef,
     iterations: number,
     maxTries: number,
-    useOld: boolean = false,
   ) {
-    let miniSrc: TileRef | TileRef[];
-
-    if (Array.isArray(src)) {
-      // Handle multiple source points
-      miniSrc = src.map((srcPoint) =>
-        this.miniMap.ref(
-          Math.floor(gameMap.x(srcPoint) / 2),
-          Math.floor(gameMap.y(srcPoint) / 2),
-        ),
-      );
-    } else {
-      // Handle single source point
-      miniSrc = this.miniMap.ref(
-        Math.floor(gameMap.x(src) / 2),
-        Math.floor(gameMap.y(src) / 2),
-      );
-    }
+    const srcArray: TileRef[] = Array.isArray(src) ? src : [src];
+    const miniSrc = srcArray.map((srcPoint) =>
+      this.miniMap.ref(
+        Math.floor(gameMap.x(srcPoint) / 2),
+        Math.floor(gameMap.y(srcPoint) / 2),
+      ),
+    );
 
     const miniDst = this.miniMap.ref(
       Math.floor(gameMap.x(dst) / 2),
       Math.floor(gameMap.y(dst) / 2),
     );
 
-    this.aStar = useOld
-      ? new SerialAStarOld(
-          miniSrc as TileRef,
-          miniDst,
-          iterations,
-          maxTries,
-          this.miniMap,
-        )
-      : new SerialAStar(miniSrc, miniDst, iterations, maxTries, this.miniMap);
+    this.aStar = new SerialAStar(
+      miniSrc,
+      miniDst,
+      iterations,
+      maxTries,
+      this.miniMap,
+    );
   }
 
   compute(): PathFindResultType {
